fix: role composition memo key must be per-target, not role-global or unguarded
A role's deferred (non-declaration) body must run exactly once per (role,
composed target) -- not once per role globally, and not once per every
composition call. Built a case table against real raku (2026.06) covering
three composition paths (class-header does, value but/does, runtime
does-on-Instance) and found mutsu diverged in five of seven scenarios:

- The class-composition path (register_class_decl -> compose_role_into_class)
  had no guard at all: re-declaring the same class (e.g. inside a loop)
  re-ran the role's deferred body on every pass.
- The value-mixin path's `mixin:{role}` memo key was role-global, ignoring
  the base type, so `1 but R` then `"x" but R` wrongly ran the body once
  instead of twice (Rakudo memoizes the composed TYPE -- Int+{R} vs Str+{R}
  are different types).
- does_rebless_instance (does on an Instance) called both a standalone
  run_mixin_role_body AND ensure_mixin_class's class-composition path,
  double-running the deferred body for a single `does` call.

Fixed all three: a (class_name, role) memo key in compose_role_into_class
(with error rollback so a dying parameterization guard still rejects every
retry, and a hoisted-shell exemption so the throwaway forward-reference
shell registration doesn't starve or double-run the real declaration's
composition), a base-type-qualified mixin key in compose_role_on_value, and
removal of the redundant run_mixin_role_body call.

Split registration_class_compose.rs's record_class_composed_roles into its
own module to stay under the 500-line file convention after these changes.

Full case table, the raku quirks found while building it (compile-time
class-header composition, literal-value but constant-folding), and the two
regressions caught along the way (role-body-guard-parameterisation.t's
retry-after-death, run-nested-role-body.t's hoisted-shell double count) are
in news/2026-08/role-composition-memo-key-raku-case-table.md.

Resolves todo/deep/adr0019-role-composition-memo-guard-raku-case-table.md.
