+++ src/runtime/registration_class_body_does.rs
@@ -83,6 +83,7 @@ impl Interpreter {
             class_lang_rev: cx.class_lang_rev,
             class_def: &mut cx.class_def,
             out: RoleCompositionOutcome::default(),
+            is_hoisted_shell: cx.is_hoisted_shell,
         };
         self.compose_role_into_class(
             &mut composition,
